Backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pytest
pytest-cov
httpx